Ajouts droits accèes tutoriel version 1

+++ src/Web/Pages/Tutoriel/Consultation.cshtml.cs
@@ -1,5 +1,6 @@
 ﻿using Gwenael.Domain;
 using Gwenael.Domain.Entities;
+using Gwenael.Web.FctUtils;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Spk.Common.Helpers.String;
@@ -23,14 +24,28 @@ namespace Gwenael.Web.Pages
             public Tutos tutoriel { get; set; }
             public List<RangeeTutos> lstRangeeTuto { get; set; }
             public string id { get; set; }
+            public bool droitAccess { get; set; }
         }
         public ConsultationModel(GwenaelDbContext pDb)
         {
             _db = pDb;
         }
 
+        public Guid ObtenirIdDuUserSelonEmail(string email)
+        {
+            User user = (User)_db.Users.Where(u => u.UserName == email).First();
+            return user.Id;
+        }
         public IActionResult OnGet()
         {
+            if (User.Identity.IsAuthenticated)
+            {
+                Guid idConnectedUser = ObtenirIdDuUserSelonEmail(User.Identity.Name);
+                if (Permission.EstGestionnaireDeContenu(idConnectedUser, _db))
+                    Input.droitAccess = true;
+                else Input.droitAccess = false;
+            } else Input.droitAccess = false;
+
             Input = new InputModel();
 
             if (Request.Query.Count >= 1)
@@ -59,10 +74,19 @@ namespace Gwenael.Web.Pages
         {
             try
             {
-                _db.Tutos.Remove(entity: _db.Tutos.Where(t => t.Id == Guid.Parse(tutoVal.tutorielIdVal)).First());
-                _db.SaveChanges();
+                if (User.Identity.IsAuthenticated)
+                {
+                    Guid idConnectedUser = ObtenirIdDuUserSelonEmail(User.Identity.Name);
+                    if (Permission.EstGestionnaireDeContenu(idConnectedUser, _db))
+                    {
+                        _db.Tutos.Remove(entity: _db.Tutos.Where(t => t.Id == Guid.Parse(tutoVal.tutorielIdVal)).First());
+                        _db.SaveChanges();
 
-                return Redirect("/tutoriel?deleteStatus=true");
+                        return Redirect("/tutoriel?deleteStatus=true");
+                    }
+                }
+                return Redirect("/tutoriel");
+
             }
             catch (Exception)
             {
@@ -74,10 +98,18 @@ namespace Gwenael.Web.Pages
         {
             try
             {
-                _db.Tutos.Where(t => t.Id == Guid.Parse(tutoVal.tutorielIdVal)).First().EstPublier = false;
-                _db.SaveChanges();
+                if (User.Identity.IsAuthenticated)
+                {
+                    Guid idConnectedUser = ObtenirIdDuUserSelonEmail(User.Identity.Name);
+                    if (Permission.EstGestionnaireDeContenu(idConnectedUser, _db))
+                    {
+                        _db.Tutos.Where(t => t.Id == Guid.Parse(tutoVal.tutorielIdVal)).First().EstPublier = false;
+                        _db.SaveChanges();
 
-                return Redirect("/tutoriel?unPublishStatus=true");
+                        return Redirect("/tutoriel?unPublishStatus=true");
+                    }
+                }
+                return Redirect("/tutoriel");
             }
             catch (Exception)
             {
